+++ src/components/radio-button/radio-button.stories.mdx
@@ -58,13 +58,6 @@ A legend can be set for the group, and each radio button can have a label.
   <Story name="with legend and labels" story={stories.WithLegendAndLabels} />
 </Canvas>
 
-### Required
-
-You can use the `required` prop to indicate if the field is mandatory.
-
-<Canvas>
-  <Story name="required" story={stories.Required} />
-</Canvas>
 
 ### With inline legend
 
@@ -138,34 +131,6 @@ Passing in `large` for the `size` prop on each radio button will increase the si
 
 <Canvas>
   <Story name="with large radio buttons" story={stories.WithLargeRadioButtons} />
-</Canvas>
-
-### With validations on radio buttons
-
-Validations can be added to the radio buttons as strings or booleans.
-
-<Canvas>
-  <Story name="with validations on buttons" story={stories.WithValidationsOnButtons} />
-</Canvas>
-
-It is possible to use the `tooltipPosition` to override the default placement of tooltips rendered as part of this component.
-
-<Canvas>
-  <Story name="with validations on buttons - tooltipPosition overriden" story={stories.WithTooltipPosition} />
-</Canvas>
-
-### With validations on radio group
-
-Validations can be added to the radio group as a string or boolean.
-
-<Canvas>
-  <Story name="with validations on group" story={stories.WithValidationsOnRadioGroup} />
-</Canvas>
-
-It is possible to use the `tooltipPosition` to override the default placement of tooltips rendered as part of this component.
-
-<Canvas>
-  <Story name="with validations on group - tooltipPosition overriden" story={stories.WithTooltipPositionOnRadioGroup} />
 </Canvas>
 
 ### With custom styled labels
